Module ovrt_druid_example::sink

This organizes how ovrt-sys passes events/callbacks into this druid app instance. But in short, this is boilerplate.

The function submit_cmd is passed to ovrt_sys, which forwards events/callbacks into that function.
The function, on ther other hand, need access into the ExtEventSink, which is always available as a static instance defined in this module (see druid_global_event_sink).

See ovrt_sys::cmd::sink and druid's blocking functions for more information.

Structs

Delegate

Functions

druid_global_event_sink
This function is used do access a static item that is able to insert arbitrary commands/events into the Druid data/control flow.
One of such types of commands is the cmd::OVRT_COMMAND, which will represent all kinds of events/callbacks from ovrt-sys.
Check cmd and ovrt_sys::cmd for more information.